Add single tournament view

+++ migrations/V240102_1__Create_table_tournaments.sql
@@ -2,6 +2,6 @@ CREATE TABLE tournaments (
     id TEXT NOT NULL PRIMARY KEY,
     url TEXT NOT NULL,
     name TEXT NOT NULL,
-    start_date BIGINT NOT NULL,
-    end_date BIGINT NOT NULL
+    start_date INTEGER NOT NULL,
+    end_date INTEGER NOT NULL
 );
